requirements.txt
requests==2.32.3
python-dotenv==1.0.1
python-telegram-bot==21.10
notion-client==2.3.0
